increases database integers storage size to 64bit values

+++ scripts/mysql/01_event_streams_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `event_streams` (
-  `no` INT(11) NOT NULL AUTO_INCREMENT,
+  `no` BIGINT(20) NOT NULL AUTO_INCREMENT,
   `real_stream_name` VARCHAR(150) NOT NULL,
   `stream_name` CHAR(41) NOT NULL,
   `metadata` JSON,
